changing modules to nest content, quizzes and exercises inside dashboard component
