+++ ScrumPlatform/public/src/app/dashboard/about/about.component.html
