+++ ScrumPlatform/public/src/app/dashboard/about/about.component.spec.ts
