+++ ScrumPlatform/public/src/app/dashboard/about/about.component.ts
@@ -3,11 +3,11 @@ import { ActivatedRoute, Params, Router } from '@angular/router';
 import { HttpService } from '../../http.service';
 
 @Component({
-  selector: 'app-daily-scrum',
-  templateUrl: './daily-scrum.component.html',
-  styleUrls: ['./daily-scrum.component.css']
+  selector: 'app-about',
+  templateUrl: './about.component.html',
+  styleUrls: ['./about.component.css']
 })
-export class DailyScrumComponent implements OnInit {
+export class AboutComponent implements OnInit {
 
   constructor(private _route: ActivatedRoute, private _httpService: HttpService, private _router: Router) { }
