+++ ScrumPlatform/public/src/app/dashboard/content/backlog-refinement/backlog-refinement.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { HttpService } from '../../http.service';
+import { HttpService } from '../../../http.service';
 
 @Component({
-  selector: 'app-dev-team',
-  templateUrl: './dev-team.component.html',
-  styleUrls: ['./dev-team.component.css']
+  selector: 'app-backlog-refinement',
+  templateUrl: './backlog-refinement.component.html',
+  styleUrls: ['./backlog-refinement.component.css']
 })
-export class DevTeamComponent implements OnInit {
+export class BacklogRefinementComponent implements OnInit {
 
   constructor(private _route: ActivatedRoute, private _httpService: HttpService, private _router: Router) { }
