+++ ScrumPlatform/public/src/app/dashboard/content/content.component.html
