+++ ScrumPlatform/public/src/app/dashboard/content/content.component.spec.ts
