+++ ScrumPlatform/public/src/app/dashboard/content/daily-scrum/daily-scrum.component.spec.ts
