+++ ScrumPlatform/public/src/app/dashboard/content/daily-scrum/daily-scrum.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { HttpService } from '../../http.service';
+import { HttpService } from '../../../http.service';
 
 @Component({
-  selector: 'app-sprint',
-  templateUrl: './sprint.component.html',
-  styleUrls: ['./sprint.component.css']
+  selector: 'app-daily-scrum',
+  templateUrl: './daily-scrum.component.html',
+  styleUrls: ['./daily-scrum.component.css']
 })
-export class SprintComponent implements OnInit {
+export class DailyScrumComponent implements OnInit {
 
   constructor(private _route: ActivatedRoute, private _httpService: HttpService, private _router: Router) { }
