+++ ScrumPlatform/public/src/app/dashboard/content/dev-team/dev-team.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { HttpService } from '../../http.service';
+import { HttpService } from '../../../http.service';
 
 @Component({
-  selector: 'app-overview',
-  templateUrl: './overview.component.html',
-  styleUrls: ['./overview.component.css']
+  selector: 'app-dev-team',
+  templateUrl: './dev-team.component.html',
+  styleUrls: ['./dev-team.component.css']
 })
-export class OverviewComponent implements OnInit {
+export class DevTeamComponent implements OnInit {
 
   constructor(private _route: ActivatedRoute, private _httpService: HttpService, private _router: Router) { }
