+++ ScrumPlatform/public/src/app/dashboard/content/overview/overview.component.html
