+++ ScrumPlatform/public/src/app/dashboard/content/overview/overview.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { HttpService } from '../../http.service';
+import { HttpService } from '../../../http.service';
 
 @Component({
-  selector: 'app-scrum-master',
-  templateUrl: './scrum-master.component.html',
-  styleUrls: ['./scrum-master.component.css']
+  selector: 'app-overview',
+  templateUrl: './overview.component.html',
+  styleUrls: ['./overview.component.css']
 })
-export class ScrumMasterComponent implements OnInit {
+export class OverviewComponent implements OnInit {
 
   constructor(private _route: ActivatedRoute, private _httpService: HttpService, private _router: Router) { }
