+++ ScrumPlatform/public/src/app/dashboard/content/product-owner/product-owner.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { HttpService } from '../../http.service';
+import { HttpService } from '../../../http.service';
 
 @Component({
-  selector: 'app-backlog-refinement',
-  templateUrl: './backlog-refinement.component.html',
-  styleUrls: ['./backlog-refinement.component.css']
+  selector: 'app-product-owner',
+  templateUrl: './product-owner.component.html',
+  styleUrls: ['./product-owner.component.css']
 })
-export class BacklogRefinementComponent implements OnInit {
+export class ProductOwnerComponent implements OnInit {
 
   constructor(private _route: ActivatedRoute, private _httpService: HttpService, private _router: Router) { }
