+++ ScrumPlatform/public/src/app/dashboard/content/scrum-master/scrum-master.component.spec.ts
