+++ ScrumPlatform/public/src/app/dashboard/content/scrum-master/scrum-master.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { HttpService } from '../../http.service';
+import { HttpService } from '../../../http.service';
 
 @Component({
-  selector: 'app-sprint-retrospective',
-  templateUrl: './sprint-retrospective.component.html',
-  styleUrls: ['./sprint-retrospective.component.css']
+  selector: 'app-scrum-master',
+  templateUrl: './scrum-master.component.html',
+  styleUrls: ['./scrum-master.component.css']
 })
-export class SprintRetrospectiveComponent implements OnInit {
+export class ScrumMasterComponent implements OnInit {
 
   constructor(private _route: ActivatedRoute, private _httpService: HttpService, private _router: Router) { }
