+++ ScrumPlatform/public/src/app/dashboard/content/sprint-cycle/sprint-cycle.component.spec.ts
