+++ ScrumPlatform/public/src/app/dashboard/content/sprint-cycle/sprint-cycle.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { HttpService } from '../http.service';
+import { HttpService } from '../../../http.service';
 
 @Component({
-  selector: 'app-about',
-  templateUrl: './about.component.html',
-  styleUrls: ['./about.component.css']
+  selector: 'app-sprint-cycle',
+  templateUrl: './sprint-cycle.component.html',
+  styleUrls: ['./sprint-cycle.component.css']
 })
-export class AboutComponent implements OnInit {
+export class SprintCycleComponent implements OnInit {
 
   constructor(private _route: ActivatedRoute, private _httpService: HttpService, private _router: Router) { }
