+++ ScrumPlatform/public/src/app/dashboard/content/sprint-planning/sprint-planning.component.html
