+++ ScrumPlatform/public/src/app/dashboard/content/sprint-retrospective/sprint-retrospective.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { HttpService } from '../../http.service';
+import { HttpService } from '../../../http.service';
 
 @Component({
-  selector: 'app-sprint-review',
-  templateUrl: './sprint-review.component.html',
-  styleUrls: ['./sprint-review.component.css']
+  selector: 'app-sprint-retrospective',
+  templateUrl: './sprint-retrospective.component.html',
+  styleUrls: ['./sprint-retrospective.component.css']
 })
-export class SprintReviewComponent implements OnInit {
+export class SprintRetrospectiveComponent implements OnInit {
 
   constructor(private _route: ActivatedRoute, private _httpService: HttpService, private _router: Router) { }
