+++ ScrumPlatform/public/src/app/dashboard/content/sprint-review/sprint-review.component.html
