+++ ScrumPlatform/public/src/app/dashboard/content/sprint-review/sprint-review.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { HttpService } from '../../http.service';
+import { HttpService } from '../../../http.service';
 
 @Component({
-  selector: 'app-sprint-planning',
-  templateUrl: './sprint-planning.component.html',
-  styleUrls: ['./sprint-planning.component.css']
+  selector: 'app-sprint-review',
+  templateUrl: './sprint-review.component.html',
+  styleUrls: ['./sprint-review.component.css']
 })
-export class SprintPlanningComponent implements OnInit {
+export class SprintReviewComponent implements OnInit {
 
   constructor(private _route: ActivatedRoute, private _httpService: HttpService, private _router: Router) { }
