+++ ScrumPlatform/public/src/app/dashboard/content/sprint/sprint.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { HttpService } from '../../http.service';
+import { HttpService } from '../../../http.service';
 
 @Component({
-  selector: 'app-product-owner',
-  templateUrl: './product-owner.component.html',
-  styleUrls: ['./product-owner.component.css']
+  selector: 'app-sprint',
+  templateUrl: './sprint.component.html',
+  styleUrls: ['./sprint.component.css']
 })
-export class ProductOwnerComponent implements OnInit {
+export class SprintComponent implements OnInit {
 
   constructor(private _route: ActivatedRoute, private _httpService: HttpService, private _router: Router) { }
