+++ ScrumPlatform/public/src/app/dashboard/content/user-stories/user-stories.component.html
