+++ ScrumPlatform/public/src/app/dashboard/dashboard.component.html
@@ -1,0 +1,4 @@
+<p>
+  dashboard works!
+</p>
+<router-outlet></router-outlet>
