+++ ScrumPlatform/public/src/app/dashboard/dashboard.component.ts
@@ -1,0 +1,17 @@
+import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute, Params, Router } from '@angular/router';
+import { HttpService } from '../http.service';
+
+@Component({
+  selector: 'app-dashboard',
+  templateUrl: './dashboard.component.html',
+  styleUrls: ['./dashboard.component.css']
+})
+export class DashboardComponent implements OnInit {
+
+  constructor(private _route: ActivatedRoute, private _httpService: HttpService, private _router: Router) { }
+
+  ngOnInit() {
+  }
+
+}
